refactor: Company function names clearer
