+++ src/api-gateway/src/company/tests/company.service.spec.ts
@@ -1,13 +1,14 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { CompanyController } from './company.controller';
-import { UsersService } from '../users/users.service';
-import { userStub } from '../../test/stubs/user.stub';
+import { CompanyService } from '../company.service';
+import { UsersService } from '../../users/users.service';
+import { userStub } from '../../../test/stubs/user.stub';
 import { MockFunctionMetadata, ModuleMocker } from 'jest-mock';
+import { CompanyController } from '../company.controller';
 
 const moduleMocker = new ModuleMocker(global);
 
-describe('CompanyController', () => {
-  let controller: CompanyController;
+describe('CompanyService', () => {
+  let service: CompanyService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
@@ -33,10 +34,10 @@ describe('CompanyController', () => {
       })
       .compile();
 
-    controller = module.get<CompanyController>(CompanyController);
+    service = module.get<CompanyService>(CompanyService);
   });
 
   it('should be defined', () => {
-    expect(controller).toBeDefined();
+    expect(service).toBeDefined();
   });
 });
