+++ src/api-gateway/src/users/tests/users.controller.spec.ts
@@ -48,7 +48,7 @@ describe('UsersController', () => {
   });
 
   it('should find all users', function () {
-    expect(usersController.findAll()).toBeDefined();
+    expect(usersController.findAllCompanies()).toBeDefined();
   });
 });
 */
